// crates/dbt-docs-server/web/src/shared/typings/domain/asset.ts
export type ResourceType =
  | 'model'
  | 'seed'
  | 'snapshot'
  | 'source'
  | 'exposure'
  | 'metric'
  | 'macro'
  | 'semantic_model'
  | 'test'
  | 'unit_test'
  | 'saved_query'
  | 'function'
  | 'group'
  | 'analysis'
  | 'operation';